+++ puzikov.herman/T3/functional.cpp
@@ -23,17 +23,3 @@ puzikov::VertIt puzikov::minVertElement(VertIt first, VertIt last, const Vertice
 {
   return std::min_element(first, last, comp);
 }
-
-void puzikov::readPolygons(std::istream &in, std::vector< Polygon > &polygons)
-{
-  using input_it_t = std::istream_iterator< Polygon >;
-  while (!in.eof())
-  {
-    std::copy(input_it_t(in), input_it_t(), std::back_inserter(polygons));
-    if (in.fail())
-    {
-      in.clear();
-      in.ignore(std::numeric_limits< std::streamsize >::max(), '\n');
-    }
-  }
-}
